Let the description tell me what scenic is
A friend linked the hex package to me, it was hard to tell me what scenic is about other than it being the "core" of something. I don't believe my description is the best (I basically took the first sentence from the README as it seemed to most accurately convey it). Happy for another description but I think it'd help people getting to know what scenic is - I'd also propose to change the title/description at github :)

Maybe even include that it's 2D? Not sure.