Move C build products under _build
The build products were being stored in the source directory. This was
causing trouble when building for different architectures since build
products for one could be used for the other. The workaround was to
build clean when switching. `elixir_make` v0.5.0 makes it easier to move
build products under the appropriate directory under `_build` by
exporting the `MIX_COMPILE_PATH` environment variable. This updates the
`elixir_make` dep to at least v0.5.0 and uses the new environment
variables.