Format and expand 0.10.0 changelog
Move the "top 3" features/changes (in my opinion) up to the top of the list

Create separate deprecations and breaking changes sections

Expand on the `push_graph/1` change using content from the warning printed out
in `Scene/push_graph/1` to help users change to the new callback-based form. I
opted to use the explicit keyword list format as it may aid comprehension for
users that are not as familiar with Elixir's keyword lists.