Fixes typo in docs
This did not read well and looks like a typo.
I added what I thought this should be but I may be wrong
and this change may not represent the original intention.
If so let me know and I will update the PR to include the proper words.

Todd Pickell @tapickell <tapickell@gmail.com>