Change the input style from just a true/false value to a list of input types. Example: input: [:cursor_button, :cursor_pos] Adding these inputs to a graph, then pushing that graph, automatically informs the drivers that they should start sending that type of input event.
Note that api base input requests are now item independent!

So use the input style when you only want the event when the cursor is over
a specific item. use the input_request or input_capture apis when you always
want the event regardless if it is over anything or not.