Make the graph compiler independent of the ViewPort.
Previously the id of any script was a positive integer and the name was a string or pid or whatever.
This required an ETS table to map between the two in the ViewPort. It also meant the VP was involved
in assigning id numbers to any given name. The big downside is that means you can't map a graph
at compile time.

This change goes with string IDs all the way for scripts. This simplifies the tables and means a
graph can be compiled down to a script at build time.

The GLFW driver will also need to be pulled to use this change.