The bounds/1 function will compute the bounding box for a graph. It does this by walking all the primitives in the graph, accounts for any transforms (rotation, scale, translate, etc...) and returns a bounding box in the format of {left, top, right, bottom} in the local coordinate space.
This is intended for use by dynamic layout engines.

Computing the bounding box of a graph is independent of scenes and can thus be used at compile time. It does a fair bit of math and work to crawl the tree, so I don't recommend you use it in a tight loop.

At this time it accounts for all primitives except Script, which it probably won't support going forward.

Components are accounted for when calculating the bounds if they support the new Components.bounds/2 callback. All the standard components have it implemented.

This commit includes several updates
  * Initial support for finding the bounding box of a graph.
  * Moved Scenic.ViewPort.GraphCompiler to Scenic.Graph.Compiler now that it doesn't depend on the ViewPort
  * Fixed a transform (position) but in the graph compiler
  * The default pin for any primitive is now set when the primitive is created or updated in the graph.
  * Components have a new optional callback to provide the default pin. {0,0} is used if no callback.
  * Components have a new optional callback to provide a bounding box
  * All standard components support the new bounds callback
  * Use Graph.bounds/1 to compute the bounding box
  * Graph.compile/1 is now the preferred API to transform a graph into a script

Caveats
  * The Script primitive is ignored when computing a Graph's bounding box. This will probably not be fixed in the future.
  * The computed bounding box does not include primitives with hidden: true
  * The computed bounding box ignores the scissor style. This may be supported in the future.