Increase nimble_options requirement
Scenic uses `{:or, _}` which was only added in v0.3.3 (and improved in
v0.3.4). v0.3.3 might be fine as the minimum required version but
nimble_options doesn't have any breaking change and I'd rather err on
the side of caution.

The resulting restriction `>= 0.3.4 and < 0.4.0` is fairly similar to
`~> 0.3`, except it won't allow usage of 0.3.0 - 0.3.3.