Doc updates for Primitives
Doc updates for `Scenic.Primitive` and `Scenic.Primitive.Component`

Removed the link to `Scenic.Primitive.Component` from
`Scenic.Primitive`'s table because the `component/3` helper no longer
exists and `Scenic.Primitive.Component` no longer seems to match in that
table.

Use plain backticks to create links when possible because it is rendered
better in editors and is shorter with better error checking in
ex_doc (which then caught the `component/3` and `circle/3` broken
links).

Also:
- Formatting updates
- Grammar updates
- Remove some dead code
- Add more links to modules