Allows `assign_new/2` to take a map of values
Consistency in the API is essential in line with the recent work of
`assign/2` taking a list or map. Building on the previous work allows
`assign_new/2` to work with either a keyword list or map.

The change is applied to `Driver` and `Scene` so both have a consistent
API.

Thanks to @amclain and @thejohncotton for the inspiration.