Warns missing `handle_input/3` on `request_input/2`
With all the different handlers for events, inputs, and messages, it
can be simple to pick the wrong one when implementing. This leads to
bugs that are difficult to find.

I didn’t want development workflows or older projects to break, so I
thought a warning would be an excellent solution to reduce the time
spent debugging but still allow current projects to continue to work
with this enhancement. This means that it is still possible to implement
the wrong callback, but hopefully, the warning will point the developer
in the right direction and save minutes, if not hours, of debugging.